ai-news-collector/requirements.txt
feedparser>=6.0.0
requests>=2.25.0
jinja2>=3.0.0
pyyaml>=5.4.0